Reorganize ui module into report/ and import/ sub-modules
Move ui/{app,event,render}.rs into ui/report/ to mirror ui/import/.
Extract TableNav into ui/table.rs (shared by both sub-modules).
Expose report and import as pub mod so callers use ui::report::* and
ui::import::* directly.

Add ui/keys.rs with is_ctrl(KeyModifiers) helper shared by both event
loops. The import/ version already masked SHIFT before comparing;
report/ was using .contains() which would also match Ctrl+Alt. Five
unit tests cover: no modifiers, Ctrl, Shift+Ctrl, Ctrl+Alt, Shift.
