+++ cli/src/ui/report.rs
@@ -1,0 +1,39 @@
+//! Report TUI session — balance and register viewer.
+//!
+//! Architecture (Elm-style; see
+//! <https://ratatui.rs/concepts/application-patterns/the-elm-architecture/>):
+//! - [`app::App`] holds all state and exposes [`app::App::update`] for transitions.
+//! - [`event`] translates `KeyEvent`s to messages and runs the loop.
+//! - [`render`] is a pure view over `App`.
+//!
+//! See [`run_ui`] for the public entry point.
+
+mod app;
+mod event;
+mod render;
+
+pub use app::{App, BalanceRow, RegisterQueryTemplate};
+
+use okane_core::report::ReportContext;
+use okane_core::report::query::Ledger;
+
+/// Runs the TUI session with the prepared [`App`].
+///
+/// Sets up the terminal (raw mode, alternate screen), installs a panic hook
+/// that restores the terminal before propagating, runs the event loop, and
+/// tears down the terminal on exit (normal or error).
+///
+/// `ledger` is borrowed mutably so the register view can be computed on
+/// demand. `ctx` is borrowed for the lifetime of the session so amount
+/// values can be formatted lazily under the same `ReportContext` that
+/// produced them.
+pub fn run_ui<'ctx>(
+    mut app: App<'ctx>,
+    ledger: &mut Ledger<'ctx>,
+    ctx: &ReportContext<'ctx>,
+) -> anyhow::Result<()> {
+    let mut terminal = ratatui::init();
+    let result = event::run(&mut terminal, &mut app, ledger, ctx);
+    ratatui::restore();
+    result
+}
